Allow to use minishift >= v1.13.0
minishift v.1.13.0 change storage driver for docker from device mapper
to overlay2. And therefore it hit a bug with wrong SELinux label of files
in /var/lib/docker

This is a workaround for images which does not have a fix
https://github.com/minishift/minishift-centos-iso/issues/216
https://github.com/minishift/minishift-centos-iso/pull/217

This commit might be reverted after some time when
standard/default images would contain correct SELinux labeling.