No need to restart the VM if the config was not updated
In order to apply the cpu mode="host-passthrough" config to the VM, one needs
to stop the VM first. That was the case before, every time. However, at
least on Fedora, the cpu mode="host-passthrough" is configured by
default and thus there is no need to actually update the config and then
restart the VM. This considerably speeds up the the process.

Note about the setup_minishift and start_minishift flag values - there
are four possible cases:
  1) 'setup_minishift == true and start_minishift == true' --- this will
     result in creating and initializing brand new minishift profile, which
     will be running after the init is done.

  2) 'setup_minishift == true and start_minishift == false' --- this will
     result in creating and initializing brand new minishift profile,
     but it will be stopped at the end (as start_minishift == false).
     It may well be started again by e.g. the os_temps role which
     usually comes after the minishift role.

  3) 'setup_minishift == false and start_minishift == true' --- this
     won't initialize (create) the VM at all. The VM is expected to be
     created before running contra-env-setup. The VM will, however, be
     started at the end (in case it's not already running).

  4) 'setup_minishift == false and start_minishift == false' --- this
     won't do minishift operations at all. This combination is useful
     for deploying into remote OpenShift clusters.